equirect2cubemap: Make it possible to have the outermost texels of the texture correspond to the edge of the cube
This in particular means that the lines of texels on the edge of the
face image will repeat in both neighbors of the corresponding edge of
the cube. This lets us have "manual" cube mapping, i.e. implemented by
texture coordinates of object vertices instead of direct support like
GL_ARB_texture_cube_map.